+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # Change Log
 ---
+
+- [feat: support lossless register and deregister](https://github.com/Tencent/spring-cloud-tencent/issues/977)